pyproject.toml
[project]
name = "event-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "fastapi[all]>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "starlette-admin[sqlalchemy]>=0.15.1",
]
